Add NewTable method for creating a new Table structure.
There was previously no easy way to create an entirely new Table.
The assumption was mostly that some other software had written
a GPT table and this software would read it.

NewTable allows the user to create a new gpt.Table that can be
written to disk with with gpt.Table.Write().