Add NewTable method for creating a new Table structure. (#6)
* Add NewTable method for creating a new Table structure.

There was previously no easy way to create an entirely new Table.
The assumption was mostly that some other software had written
a GPT table and this software would read it.

NewTable allows the user to create a new gpt.Table that can be
written to disk with with gpt.Table.Write().

* Adjust call signature of NewTable to take a NewTableArgs.

In order to allow for adding new args in the future without
losing compatibility, change NewTable to take a NewTableArgs
type.  If values in NewTableArgs are un-set, then defaults
will be used, including creation of a new UUID.

* Fix calculation of first usable LBA in NewTable.

Previously the value 34 was just hard coded as the first usable LBA.
That is based on a sector size of 512.  The partitions fit into 5
sectors if the sector size is 4096.  Actually calculate rather than
hard coding.